Removed timer-based rendering, added config option
Previously the timeline manager had a tick() function that was checking
if the media.currentTime was roughly close to a CEA-708 message PTS
value, and feeding them into to CEA-708 interpreter, which mapped these
back to media.currentTime.

This had no value, and only wasted CPU, since the video tag allows
creation of all the cues in advance, with the start/end times on them.

Additionally, i added an enableCEA708Captions boolean in hls.config

Lastly, updated the interpreter to support more than one caption in the
buffer before displaying it (which is required by the spec, but not
used by my sample media, currently)