mediaController : implement exponential backoff retry
similar to the one performed in xhr-loader.
this logic has been moved to mediaController to have a better control of the load retry mechanism.
in case of load error/timeout,
we first try to switch to redundant,
then to switch down to level 0,
then we retry several times at level 0
if we keep the logic in xhr-loader we cannot switch to redundant stream straight-away.