ts-demuxer: ensure we have enough bytes to read NAL unit type in case NALus are overlapping
when NALus are overlapping accross PES packets,
it can happen that the last 3 bytes of a PES packet are the NALu start code, and that the NAL unit type is on the following PES packet.
that case was not handled properly. ensure we have enough bytes to read NALu type ...

related to https://github.com/dailymotion/hls.js/issues/176