Optimizations and renaming variables
Removes more DataViews and does more work in a single loop iteration.  Increases speed by manipulating data 32 bits at a time instead of doing a lot of shifting using 8 bits at a time.  This also reduces the number of iterations we do when processing data.  Also comment the code and changes variable names so it's easier to read.

JW7-3391