Playback streams without ID3 data or ID3 timestamp data
Allows playback streams which do not include ID3 data by setting a default starting location for searching through bytes of AAC data at 0 instead of allowing an undefined value to be used.
Allows playback of streams which do not include ID3 timestamp data by generating a starting PTS value from the timeOffset provided when demuxing audio data.
Improved the mask we're using to search for the ID3 data header as it can be more specific than what it is right now.

JW7-3382