Use ID3 Timestamp information only when parsing alternate audio renditions
In order to support AAC audio only streams and v4 audio tracks, the AACDemuxer can get PTS for segments in two different ways.  When playing back v4 audio streams, we need to use ID3 data to set the PTS of the segments correctly.  In this situation, the AADDemxuer is initialized with an id of 'audio' and therefore use the timeStamp from the ID3 data to set the PTS.  Otherwise, we want to ignore the timestamp value because when playing back a single audio track, we expect to be able to append audio segments one after another in order to get the expected playback.

JW7-3452