Create maps to translate PTS values and Discontinuity Sequence Ids to align content
This change creates maps that record what timeline time combinations of Discontinuity Sequence Ids and PTS value in order to let content write to the expected location in the stream.  This fixes issues when switching between quality levels in live streams which contain discontinuities by allowing the remuxer to successfully shift the information parsed out of a segment its position relative to the previous stream.  This allows stream to set the information that's stored about the level to have the correct time offset used by the stream controller to load subsequent fragments.  Without this, a newly loaded stream is assumed to start at a time of 0, and thus will stall for a very long time until the manifest loads enough segments that it will reach the current time in the video player.

JW7-3199