Don't set a default value for positionAlign
Don't set a default value for positionAlign when parsing VTT files. FF is the only browser that supports this property in its VTTCue interface and defaults to center. This value is of no use in Chrome or Safari, where it's not supported. In IE/Edge, where we polyfill the VTTCue interface and fallback to TextTrackCue, this value breaks the parser.