Enhance ability to detect when short fragments must be downloaded
These changes enhance the ability of the stream controllers to know when short segments need to be downloaded by ensuring that the maxFragLookUpTolerance is at least as large as the smallest segment.  If the value is larger than the smallest segment, then the binary search can pass over segments necessary to playback.
This also makes it easier to try find the correct fragment to download more quickly by checking the item with the next sn value to see if it should be used before searching all fragments that are available first.

JW7-3621