level-helper: update frag with PTS in fragments array
in case a live playlist has been refreshed between frag loading and updateFragPTSDTS(), fragments array will not contain the initial frag object.
As updateFragPTSDTS() is doing the PTS update on the original frag object, we need to update fragments array with this original frag, so that PTS info could be propagated in the playlist.
this to avoid invalid sliding computation

related to #1172