Avoid polluting window object
By assigning values to window.MediaSource and window.SourceBuffer
we create properties on the window object which did not necessarily
exist before. On iPhones, for example, window.MediaSource does not
exist, but when we do window.MediaSource = window.MediaSource || window.WebKitMediaSource
we declare the property with an undefined value. This changes the
outcome of an (arguably) common method for checking whether the
browser supports a certain Web API, namely by using the "in" operator
like so:

if ('MediaSource' in window) { ... }

The property declaration was introduced in 66ed87a7

This commit lets window.MediaSource and window.SourceBuffer remain
undeclared on the window object on browsers which do not support
MediaSource/SourceBuffer.