Adjust webvtt cue offsets with mpegts rollovers
This is a fix for live streams where the mpegts clock wraps multiple
times during a long running stream. The `rolloverOffset` is based on the
start time of the first cue which is assumed to be absolute to the
stream start. It's unclear if this fix supports webvtt discontinuities
and other special cases, but it handles the basic case where
X-TIMESTAMP-MAP is unset, or set to LOCAL=0, MPEGTS=0.