fix: invalid array access modifier in subtitle-stream-controller
if during a subtitle playlist loading, but before it finished loading, the selected language is changed to none, e.g. hls.subtitleTrack = -1 then the subtitle-stream-controller doTick is still called when the manifest of the previous language is finished loading.
The doTick gets called from the playlist loader class, but the currentTrackId in subtitle-stream-controller is now set to -1 and the doTick method tries to access a nested array member with the -1 index and throws an error which then renders subtitle selection.

This is where it throws, so need to add protection that we don't enter if `trackId === -1'

fixes https://github.com/video-dev/hls.js/issues/1948