Config typescript (#2181)
Adds typing to the default HlsConfig.

Additionally provides controller-based config types to use for type narrowing controller configuration to its own subset of configuration properties from the exposed public interface.

Added Webpack DefinePlugin constants in their own declaration file to be used in other files as well.

A test stream from #1909 started to 404, reached out to author of issue to see if we could move this stream somewhere for continued use.

Esdoc plugin was required so that documentation could continue to be produced. The AST parser took issue with the object rest spreading.