fix: calling detatchMedia() followed by attachMedia() causes audio to not play
On re-attch the context of altAudio is not saved in buffer control so it doesn't know to create two source buffers.
Also, in context of stream controller it doesn't reset the `altAudio` flag.
I also noticed that stream controller is reloading the entire frags it has in `fragmentTracker` and it seems it needs to be reset on detach, otherwise there's a very long start delay where unncessery buffering occurs.

based on #2110 and fixes #2099.