gap-controller: add handling for audio-only autoplay issues in Chrome
on ultimate playhead-nudge retry, we can assume there may be a real issue
with the browser not allowing playback for another reason than some of the
situations we could have resolved by the first attempts to fix. now we can
try and handle a real situation happening on Chrome with some audio-only content:
it seems that the playback is "blocked" there because of something related
to auto-play policies. the only programmtic way to resolve this (we can always
fix this with a real UI event calling play or using browser/device controls)
has been found to set the media element to muted state. In this case a programmatic
play() will work (as opposed to when the media is not muted, where the browser
will seem to ignore the intent plainly, eventually waiting for more user action)