audio stream: prevent stuck player by removing non-current fragments
Observed with a separate audio stream: After seeking and/or level
adjustments, the audio stream controller would fetch a future fragment
which would later be found as having FragmentState.OK, causing doTick()
to be stuck in State.IDLE - iow waiting for an already fetched fragment,
which would never be pushed to the demuxer.

This feels more like a bandaid, because I have not fully understood the
reason for the wrong fragment to be selected in the first place, but at
any rate, IIUC, the state machine can only process one fragment at a
time and requires onFragLoaded to be called in order to push data.