chore(tests): Add encrypted tests to extend DRM support.
When fragments are signalled from the manifest they are encrypted we
need to be able to detect this fact. Currently we use the boolean
"encrypted" on Fragment to support this. So I've added some tests that
cover the current behaviour for AES-128 encryption, and some skipped
tests that cover the future usecase.

Additionally, this change includes a tiny change to the LevelKey class.
It now has a private constructor, and two static methods to create
itself. `fromURL` and `fromURI`, one supports the current usecase for
absolute or relative URLs, and the other is for supporting the future
use-case of URI keys.