switch from canary to alpha always base new version of last git tag
even when last git tag is a prerelease

Switching to alpha instead of canary because the semver sorting algorithm is alphabetic and the automated release should always have a lower precendence than manual ones.

Always prefix to the last version even when it's a prelease so that the ordering is correct, and so that automated alpha releases are able to take precendence over manual ones when new commits are pushed.

See discussion in #2822