switch to alpha instead of canary and ensure alphas have lower precedense than manual tags
Switching to alpha instead of canary because the semver sorting algorithm is alphabetic and the automated release should always have a lower precendence than manual ones.

Always prefix to the last version even when it's a prerelease so that the ordering is correct, and so that automated alpha releases are able to take precendence over manual ones when new commits are pushed.

Make the prefix `0.alpha` instead of just `alpha` so that it's easy to check if a manual version could be lower than an existing alpha version without dissalowing things like `-rc.0`

See discussion in #2822