chore(deps): Upgrading TypeScript to 4.0.3.
This caused an issue with ESLint due to the change in the TypeScript
AST. Which necessitated an upgrade to the ESLint plugin to fix.

See: https://github.com/microsoft/TypeScript/issues/40446.

As this was a decently large group in some of the underlying versions
for these repos, there were a few issues that came up when I upgraded
with the default settings, but I went through and fixed them.

Some of it was questionable because of our use of a bundler which will
take care of some of the "no-use-before-define" rule. But, I left the
defaults to try and stay as similar to the rest of the ecosystem as
possible.