Fix bandwidth sampling for small transfers and fast connections
The estimator factored the transfer duration into the weight. This means that
very fast transfers will have negligible influence on the measured average.

The minDelayMs_ was used to cap this effect, but it also capped the bitrate.
This meant that a 1000 byte transfer, transferred in 10ms, would register at a
rate of 160Kbps instead of 800Kbps (on top of the low weight).

The fix is to use the duration of the transferred fragment / part for the weight.
This preserves the effect of samples that match the line rate, while
significantly increasing the effect of fast transfers (due to small size, or
from a fast connection).

With 5s segments of size 100,000:
Start 10x 5 sec transfers => 160Kbps (reference)

Rate change (old):
1x 0.5 sec transfer (1.6Mbps) => 216Kbps
vs
1x 0.05 sec transfer (16Mbps) => 222Kbps (10x faster is only 4% more!!)

Rate change (new):
1x 0.5 sec transfer (1.6Mbps) => 627Kbps
vs
1x 0.05 sec transfer (16Mbps) => 5.3Mbps