id3: add duration to metadata samples, mimic Safari behavior on unbounded cues.
Related to video-dev#4964

On Safari, when an ID3 track is present, cues are initially created with
an Infinite endTime for MPEG-TS and Packed Audio sources. With fragmented
MP4, they may have an Infinite endTime if the duration is 0xffffffff, or
very small - otherwise it's finite.

As new ID3 tags come in, any unbounded cues have the endTime updated.

The current draft of VTTCue allows an unrestricted double for the
endTime - which allows for using Number.POSITIVE_INFINITY.
It hasn't rolled out yet, so this falls back to Number.MAX_VALUE.