Fix m3u8 does not reload if the target-duration is larger than part-hold-back, which is popular in low latency mode
Without this fix, m3u8 is reload after target-duration and cause the low latency stream buffer to be stalled, when
the total parts loaded in the buffer is lower than target-duration (which is a property for normal-latency mode).
This fix move to use partTarget as reference for m3u8 reload interval if the low latency mode is enabled.